moves DogGallery class to its own file
