+++ challenges/dog-photo-gallery/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dog photo gallery</title>
-    <script src="./script.js" defer></script>
+    <script type="module" src="./script.js" defer></script>
   </head>
   <body>
     <header>
